src/requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.4.1
django-coverage-plugin==1.5.0

django-extensions==1.9.6

django-debug-toolbar==1.8

# improved REPL
ipdb==0.10.3

